requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
firebase-admin==6.9.0
WeasyPrint==64.0
python-dotenv==1.1.1
psycopg2-binary==2.9.10
Jinja2==3.1.6
Pillow==11.3.0 